requirements.txt
asgiref==3.5.0
boto3==1.21.37
botocore==1.24.37
charset-normalizer==2.0.12
cryptography==36.0.2
Django==4.0.3
django-appconf==1.0.5
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-cryptography==1.1
django-filter==21.1
django-localflavor==3.1
django-mysql==4.5.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
mysql-connector-python==8.0.28
mysqlclient==2.1.0
Pillow==9.1.0
pycparser==2.21
PyMySQL==1.0.2
python-stdnum==1.17
register==0.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.0.0
